requirements.txt
Flask==3.0.0
mistune==3.0.2
python-frontmatter
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3